Add handler pipelining.
Pipelining allows an arbitrary number of renderers to get a shot
at a template before it is returned, so template.html.markdown.erb
will be passed through ERb->Markdown and then finally wrapped in a
layout. This also includes an initial refactoring of template
handling to DRY up some of the functionality, in particular the
handling of layouts.